backend/requirements.txt
fastapi[all]
sqlalchemy
bcrypt
python-jose